Start build up multi-folder structure
* Move basics about Patch into base
* Rework build process to handle multi-folder structure
* One general Patch base-class ist still missing
  --> will follow after rework of ev algorithm
