+++ cpp/BoundaryData.cpp
@@ -50,7 +50,7 @@ BoundaryData<T>::BoundaryData(
 
   std::int32_t size_fctdata = _num_rhs * _num_fcts;
   _local_fct_id.resize(_num_fcts);
-  _facet_type.resize(size_fctdata, PatchFacetType::internal);
+  _facet_type.resize(size_fctdata, base::PatchFacetType::internal);
   _offset_fctdata.resize(_num_rhs + 1);
 
   if (reconstruct_stress)
@@ -160,7 +160,7 @@ BoundaryData<T>::BoundaryData(
     for (std::int32_t fct : fct_esntbound_prime[i_rhs])
     {
       // Set facet type
-      facet_type_i[fct] = PatchFacetType::essnt_primal;
+      facet_type_i[fct] = base::PatchFacetType::essnt_primal;
     }
 
     // Handle facets with essential BCs on dual problem
@@ -334,7 +334,7 @@ BoundaryData<T>::BoundaryData(
         apply_inverse_dof_transform(boundary_values, cell_info, cell, 1);
 
         // Set values and markers
-        facet_type_i[fct] = PatchFacetType::essnt_dual;
+        facet_type_i[fct] = base::PatchFacetType::essnt_dual;
         _local_fct_id[fct] = fct_loc;
 
         for (std::size_t i = 0; i < _ndofs_per_fct; ++i)
@@ -422,7 +422,7 @@ void BoundaryData<T>::calculate_patch_bc(
     const int rhs_i, const std::int32_t fct, const std::int8_t hat_id,
     mdspan_t<const double, 2> J, const double detJ, mdspan_t<const double, 2> K)
 {
-  if (facet_type(rhs_i)[fct] == PatchFacetType::essnt_dual)
+  if (facet_type(rhs_i)[fct] == base::PatchFacetType::essnt_dual)
   {
     // The cell
     std::int32_t cell = _fct_to_cell->links(fct)[0];
