+++ cpp/base/CMakeLists.txt
@@ -1,0 +1,5 @@
+set(HEADERS_base
+    ${CMAKE_CURRENT_SOURCE_DIR}/Patch.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/deqlb_base.hpp
+    PARENT_SCOPE
+)
