+++ cpp/base/deqlb_base.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+/// @brief Miscellaneous classes, functions and types.
+///
+/// This namespace provides the basic the basic definitions of a patch, boundary
+/// conditions for equilibration problems an a local solver e.g. for
+/// projections.
+namespace dolfinx_eqlb::base
+{
+}
+
+#include "Patch.hpp"
+
+// dolfinx_eqlb common interface
