+++ cpp/dolfinx_eqlb.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+/// @brief Top-level namespace
+///
+/// A package for flux equilibration.
+namespace dolfinx_eqlb
+{
+}
+
+// dolfinx_eqlb interface
+#include "base/deqlb_base.hpp"
+#include "reconstruction.hpp"
